+++ src/pl-hash.c
@@ -20,7 +20,7 @@ alignment.
 #if WORDS_BIGENDIAN
 
 unsigned int
-MurmurHashAligned2(const void * key, int len, unsigned int seed)
+MurmurHashAligned2(const void * key, size_t len, unsigned int seed)
 { const unsigned int m = 0x5bd1e995;
   const int r = 24;
   unsigned int h = seed ^ len;
@@ -64,7 +64,7 @@ MurmurHashAligned2(const void * key, int len, unsigned int seed)
 #define MIX(h,k,m) { k *= m; k ^= k >> r; k *= m; h *= m; h ^= k; }
 
 unsigned int
-MurmurHashAligned2(const void *key, int len, unsigned int seed)
+MurmurHashAligned2(const void *key, size_t len, unsigned int seed)
 { const unsigned int m = 0x5bd1e995;
   const int r = 24;
   const unsigned char * data = (const unsigned char *)key;
