+++ src/pl-hash.h
@@ -13,6 +13,6 @@
 #define MURMUR_SEED	(0x1a3be34a)
 
 COMMON(unsigned int)
-	MurmurHashAligned2(const void *key, int len, unsigned int seed);
+	MurmurHashAligned2(const void *key, size_t len, unsigned int seed);
 
 #endif /*PL_HASH_H_INCLUDED*/
